Clean up version handling and metadata, add docs
The content of the "Broken Since" cell is now constructed automatically
based on the "product", the "firstBadVersion", and the
"lastGoodVersion".

The standard set of fields in the template bug as well as each real bug
now more precisely cover the desired information. Some of the fields
were also renamed for aesthetics.

The _template.md file now includes explanatory comments about all of the
bug metadata fields.